Simplify pull request template.
This PR removes most of the content from the pull request template. The
headers for summary and testing remain but all other content has been
removed.

In general, folks were deleting all the content as they created their
PRs so the full set of information was taking more time then it was
saving.